app/general/helpers/testify/requirements.txt
cffi==2.0.0
cryptography==46.0.2
greenlet==3.2.4
lxml==6.0.2
numpy==2.3.3
pandas==2.3.3
pip==25.1.1
pycparser==2.23
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
